+++ bullet/src/BulletDynamics/Dynamics/btDiscreteDynamicsWorld.h
@@ -4,8 +4,8 @@ Copyright (c) 2003-2009 Erwin Coumans  http://bulletphysics.org
 
 This software is provided 'as-is', without any express or implied warranty.
 In no event will the authors be held liable for any damages arising from the use of this software.
-Permission is granted to anyone to use this software for any purpose, 
-including commercial applications, and to alter it and redistribute it freely, 
+Permission is granted to anyone to use this software for any purpose,
+including commercial applications, and to alter it and redistribute it freely,
 subject to the following restrictions:
 
 1. The origin of this software must not be misrepresented; you must not claim that you wrote the original software. If you use this software in a product, an acknowledgment in the product documentation would be appreciated but is not required.
@@ -37,7 +37,7 @@ struct InplaceSolverIslandCallback;
 ATTRIBUTE_ALIGNED16(class) btDiscreteDynamicsWorld : public btDynamicsWorld
 {
 protected:
-	
+
     btAlignedObjectArray<btTypedConstraint*>	m_sortedConstraints;
 	InplaceSolverIslandCallback* 	m_solverIslandCallback;
 
@@ -62,7 +62,7 @@ protected:
 	bool	m_applySpeculativeContactRestitution;
 
 	btAlignedObjectArray<btActionInterface*>	m_actions;
-	
+
 	int	m_profileTimings;
 
 	bool	m_latencyMotionStateInterpolation;
@@ -70,13 +70,13 @@ protected:
 	btAlignedObjectArray<btPersistentManifold*>	m_predictiveManifolds;
 
 	virtual void	predictUnconstraintMotion(btScalar timeStep);
-	
+
 	virtual void	integrateTransforms(btScalar timeStep);
-		
+
 	virtual void	calculateSimulationIslands();
 
 	virtual void	solveConstraints(btContactSolverInfo& solverInfo);
-	
+
 	virtual void	updateActivationState(btScalar timeStep);
 
 	void	updateActions(btScalar timeStep);
@@ -119,13 +119,13 @@ public:
 	virtual void	addAction(btActionInterface*);
 
 	virtual void	removeAction(btActionInterface*);
-	
+
 	btSimulationIslandManager*	getSimulationIslandManager()
 	{
 		return m_islandManager;
 	}
 
-	const btSimulationIslandManager*	getSimulationIslandManager() const 
+	const btSimulationIslandManager*	getSimulationIslandManager() const
 	{
 		return m_islandManager;
 	}
@@ -158,19 +158,19 @@ public:
 	virtual void	setConstraintSolver(btConstraintSolver* solver);
 
 	virtual btConstraintSolver* getConstraintSolver();
-	
+
 	virtual	int		getNumConstraints() const;
 
 	virtual btTypedConstraint* getConstraint(int index)	;
 
 	virtual const btTypedConstraint* getConstraint(int index) const;
 
-	
+
 	virtual btDynamicsWorldType	getWorldType() const
 	{
 		return BT_DISCRETE_DYNAMICS_WORLD;
 	}
-	
+
 	///the forces on each rigidbody is accumulating together with gravity. clear this after each timestep.
 	virtual void	clearForces();
 
@@ -210,7 +210,7 @@ public:
 	{
 		m_applySpeculativeContactRestitution = enable;
 	}
-	
+
 	bool getApplySpeculativeContactRestitution() const
 	{
 		return m_applySpeculativeContactRestitution;
@@ -232,14 +232,14 @@ public:
 
 	// XXX EMSCRIPTEN: Contact callback support
 	void setContactAddedCallback(unsigned long callbackFunction) {
-		gContactAddedCallback = (ContactAddedCallback)callbackFunction;
+		gContactAddedCallback = (ContactAddedCallback)(size_t)callbackFunction;
 	}
 	void setContactProcessedCallback(unsigned long callbackFunction) {
-		gContactProcessedCallback = (ContactProcessedCallback)callbackFunction;
+		gContactProcessedCallback = (ContactProcessedCallback)(size_t)callbackFunction;
 	}
 	void setContactDestroyedCallback(unsigned long callbackFunction) {
-		gContactDestroyedCallback = (ContactDestroyedCallback)callbackFunction;
-	}	
+		gContactDestroyedCallback = (ContactDestroyedCallback)(size_t)callbackFunction;
+	}
 };
 
 #endif //BT_DISCRETE_DYNAMICS_WORLD_H
